# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/csrf-cookie": {
            "get": {
                "tags": [
                    "Get CSRF"
                ],
                "summary": "Get CSRF cookie",
                "operationId": "a9581996e229f502d0b7731e95b28ef3",
                "responses": {
                    "204": {
                        "description": "CSRF cookie set successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "4230c328926eb0a3e2f1386c0032996d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Người dùng đã đăng ký thành công. Vui lòng kiểm tra email để xác thực tài khoản của bạn!"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login account",
                "operationId": "6bc99d84bd05584feeb830297645afd9",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Người dùng đã đăng nhập thành công!"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/v1/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated user profile",
                "operationId": "1b4722d869be8b59b600bd44dd2eab79",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Sugar"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "tinhabc3009@gmail.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "https://...png"
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "is_active": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "role": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "019cd38b-7fd5-726c-8ab0-81e7c17fabc5"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "customer"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "customer"
                                                        },
                                                        "permissions": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string",
                                                                        "example": "019cfc0b-11f1-727d-a417-25a604446b3e"
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Post Management"
                                                                    },
                                                                    "key": {
                                                                        "type": "string",
                                                                        "example": "post_management"
                                                                    },
                                                                    "actions": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "properties": {
                                                                                "id": {
                                                                                    "type": "string",
                                                                                    "example": "019cfc04-566e-734e-91f6-8d2470bfba30"
                                                                                },
                                                                                "name": {
                                                                                    "type": "string",
                                                                                    "example": "Update"
                                                                                },
                                                                                "key": {
                                                                                    "type": "string",
                                                                                    "example": "update"
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout account",
                "operationId": "eddcf79a7be8ade060f7b7b8aa2e7cfd",
                "responses": {
                    "201": {
                        "description": "Đăng xuất hành công!"
                    },
                    "500": {
                        "description": "Đã có lỗi xảy ra"
                    }
                }
            }
        },
...
    },
...
}